gossip: fix lowres_clock::duration assumption
The variable diff is assigned a type of std::chrono::milliseconds
but later used to store the difference between two
lowres_clock::time_point samples. This works now because the two
types are the same, but fails if lowres_clock::duration changes.

Remove the assumption by using lowres_clock::duration.