frozen_mutation: add freeze_gently
Allow yielding in-between freezing rows.
At the moment yielding between freezing cells
or tombstones is not support yet,
assuming freezing a single row does not stall the reactor
for too long.

It may be needed to add that in the future
for supporting large (wide) rows that contain
many cells, and/or large colletions.

Refs #9866

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>