frozen_mutation: allow freeze to yield when called in a thread
Allow yielding in-between freezing rows.

At the moment yielding between freezing cells
or tombstones is not supported yet,
assuming freezing a single row does not stall the reactor
for too long.

It may be needed to add that in the future
for supporting large (wide) rows that contain
many cells, and/or large colletions.

Refs #9866

The reason we need to run in a thread is
that ser::writer_of_mutation::partition call
must be synchronous.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>